refactor: shrink readium preference realm boundary
