+++ .agents/memory-bank/activeContext.md
@@ -4,6 +4,8 @@
 The primary focus is the modernization of the EBook Reader Architecture. Reading-position persistence now has an explicit book/server boundary, and shelf bootstrap completion is deterministic. All 392 unit tests and 1 UI test pass successfully; the Mac Catalyst build also succeeds.
 
 ## Recent Changes & Decisions
+- **Book Detail Readium Nil Book Regression Fix (2026-07-03, complete):** Fixed a state mismatch where opening a book from Book Detail populated `ReadingSessionManager.readerInfo` but left `readingBook` nil or stale, which could surface as "Nil Book" when opening Readium EPUB or related reader sheets. `BookDetailViewModel` now routes read and preview-dismiss entry points through a shared helper that sets the active reading book via `CalibreBookManager`, allowing `ReadingSessionManager` to derive matching `readerInfo`. Added regression coverage to `BookDetailViewModelTests.testReadBookWhenInShelf` asserting the active reading book matches the selected detail book. Validation: focused `BookDetailViewModelTests`, `YabrReadiumReaderViewControllerTests`, `ReadiumPreferenceValueTests`, and `ReaderPreferenceRepositoryTests` passed (45 tests); `git diff --check` passed; standard iOS Simulator build passed.
+- **Readium Preference Realm Boundary Shrink (2026-07-03, complete):** Split `ReadiumPreferenceValue` out of `Views/ReadiumView/ReadiumPreferenceAdapter.swift` into `Models/Domain/ReadiumPreferenceValue.swift`, leaving the adapter responsible only for Readium/UI preference conversion. Moved `ReadiumPreferenceRealm` value mapping into `Models/Realm/ReadiumPreferenceRealmMappers.swift`, so `ReadiumView` no longer directly imports `RealmSwift` or references `ReadiumPreferenceRealm`. Repository APIs and Realm schema are unchanged. Validation: ReadiumView RealmSwift scan and adapter `ReadiumPreferenceRealm` scan passed with no matches; focused `ReadiumPreferenceValueTests`, `ReaderPreferenceRepositoryTests`, `YabrReadiumReaderViewControllerTests`, and `YabrReaderSettingsViewModelTests` passed (24 tests); standard iOS Simulator build passed.
 - **Recent Shelf Reader Presentation Regression Fix (2026-07-03, complete):** Fixed a regression where tapping a readable Recent shelf cover set the session reader presentation flag but did not immediately open the reader until another `MainViewModel` published change, such as tab switching. Root cause was the `AppContainer` observable removal: `ReadingSessionManager` state changes no longer reached `MainView` through container `objectWillChange`, while `MainView.fullScreenCover` was bound to a non-published computed property. `MainViewModel` now owns a published reader presentation flag, mirrors `ReadingSessionManager.presentingReaderSnapshots()`, and syncs dismissals back to the session manager. Added regression coverage for Recent shelf tap presentation and dismissal synchronization. Validation: focused `MainViewModelTests` and `RecentShelfViewModelTests` passed (18 tests); `git diff --check` passed; standard iOS Simulator build passed.
 - **Review Follow-up: Settings Server Status Refresh and Async Broadcaster Registration (2026-07-03, complete):** Fixed two review findings from the AppContainer/manager async migration. `ManagerAsyncBroadcaster` now registers continuations before yielding the initial value, closing a small subscription race. `SettingsViewModel` now consumes `CalibreServerManager.serverInfoStagingSnapshots()` into published view-model state, so Settings server refresh/probe status and reachability row text update even though `AppContainer` is no longer an `ObservableObject`. Added focused regression coverage for Settings server-info snapshots and broadcaster initial/subsequent delivery. Validation: focused `SettingsViewModelTests` and `CalibreServerManagerTests` passed (21 tests); `git diff --check` passed; standard iOS Simulator build passed.
 - **SwiftUI Realm Environment Injection Removal (2026-07-03, complete):** Removed the remaining `RealmSwift` import and `.environment(\.realmConfiguration, ...)` injections from `MainView.swift` and `RecentShelfView.swift`. `MainView` still gates tab rendering on database configuration readiness, but no longer binds or forwards a Realm configuration into Browse/Settings tabs; `RecentShelfView` no longer injects a per-server Realm configuration into the reading-history sheet. This is a View-layer cleanup only: persistence/repository/model RealmSwift ownership remains unchanged. Validation: target-file Realm injection scan passed with no matches; Views Realm SwiftUI dependency scan found only an unrelated commented example; `git diff --check` passed; focused `MainViewModelTests`, `RecentShelfViewModelTests`, and `BookDetailViewModelTests` passed on rerun (40 tests) after one simulator bootstrap termination; standard iOS Simulator build passed.
